Make KnownColor tests corefx-only.
