+++ src/System.Drawing.Primitives/tests/ColorTests.cs
@@ -162,62 +162,6 @@ namespace System.Drawing.Primitives.Tests
                 .Select(p => new object[] { p.Name })
                 .ToArray();
 
-        public static readonly IEnumerable<object[]> AllKnownColors = Enum.GetValues(typeof(KnownColor)).Cast<KnownColor>()
-            .Where(kc => kc != 0)
-            .Select(kc => new object[] {kc})
-            .ToArray();
-
-        public static readonly IEnumerable<object[]> SystemColors =
-            new[]
-            {
-                KnownColor.ActiveBorder, KnownColor.ActiveCaption, KnownColor.ActiveCaptionText,
-                KnownColor.AppWorkspace, KnownColor.Control, KnownColor.ControlDark, KnownColor.ControlDarkDark,
-                KnownColor.ControlLight, KnownColor.ControlLightLight, KnownColor.ControlText, KnownColor.Desktop,
-                KnownColor.GrayText, KnownColor.Highlight, KnownColor.HighlightText, KnownColor.HotTrack,
-                KnownColor.InactiveBorder, KnownColor.InactiveCaption, KnownColor.InactiveCaptionText, KnownColor.Info,
-                KnownColor.InfoText, KnownColor.Menu, KnownColor.MenuText, KnownColor.ScrollBar, KnownColor.Window,
-                KnownColor.WindowFrame, KnownColor.WindowText, KnownColor.ButtonFace, KnownColor.ButtonHighlight,
-                KnownColor.ButtonShadow, KnownColor.GradientActiveCaption, KnownColor.GradientInactiveCaption,
-                KnownColor.MenuBar, KnownColor.MenuHighlight
-            }.Select(kc => new object[] {kc}).ToArray();
-
-        public static readonly IEnumerable<object[]> NonSystemColors =
-            new[]
-            {
-                KnownColor.Transparent, KnownColor.AliceBlue, KnownColor.AntiqueWhite, KnownColor.Aqua,
-                KnownColor.Aquamarine, KnownColor.Azure, KnownColor.Beige, KnownColor.Bisque, KnownColor.Black,
-                KnownColor.BlanchedAlmond, KnownColor.Blue, KnownColor.BlueViolet, KnownColor.Brown,
-                KnownColor.BurlyWood, KnownColor.CadetBlue, KnownColor.Chartreuse, KnownColor.Chocolate,
-                KnownColor.Coral, KnownColor.CornflowerBlue, KnownColor.Cornsilk, KnownColor.Crimson, KnownColor.Cyan,
-                KnownColor.DarkBlue, KnownColor.DarkCyan, KnownColor.DarkGoldenrod, KnownColor.DarkGray,
-                KnownColor.DarkGreen, KnownColor.DarkKhaki, KnownColor.DarkMagenta, KnownColor.DarkOliveGreen,
-                KnownColor.DarkOrange, KnownColor.DarkOrchid, KnownColor.DarkRed, KnownColor.DarkSalmon,
-                KnownColor.DarkSeaGreen, KnownColor.DarkSlateBlue, KnownColor.DarkSlateGray, KnownColor.DarkTurquoise,
-                KnownColor.DarkViolet, KnownColor.DeepPink, KnownColor.DeepSkyBlue, KnownColor.DimGray,
-                KnownColor.DodgerBlue, KnownColor.Firebrick, KnownColor.FloralWhite, KnownColor.ForestGreen,
-                KnownColor.Fuchsia, KnownColor.Gainsboro, KnownColor.GhostWhite, KnownColor.Gold, KnownColor.Goldenrod,
-                KnownColor.Gray, KnownColor.Green, KnownColor.GreenYellow, KnownColor.Honeydew, KnownColor.HotPink,
-                KnownColor.IndianRed, KnownColor.Indigo, KnownColor.Ivory, KnownColor.Khaki, KnownColor.Lavender,
-                KnownColor.LavenderBlush, KnownColor.LawnGreen, KnownColor.LemonChiffon, KnownColor.LightBlue,
-                KnownColor.LightCoral, KnownColor.LightCyan, KnownColor.LightGoldenrodYellow, KnownColor.LightGray,
-                KnownColor.LightGreen, KnownColor.LightPink, KnownColor.LightSalmon, KnownColor.LightSeaGreen,
-                KnownColor.LightSkyBlue, KnownColor.LightSlateGray, KnownColor.LightSteelBlue, KnownColor.LightYellow,
-                KnownColor.Lime, KnownColor.LimeGreen, KnownColor.Linen, KnownColor.Magenta, KnownColor.Maroon,
-                KnownColor.MediumAquamarine, KnownColor.MediumBlue, KnownColor.MediumOrchid, KnownColor.MediumPurple,
-                KnownColor.MediumSeaGreen, KnownColor.MediumSlateBlue, KnownColor.MediumSpringGreen,
-                KnownColor.MediumTurquoise, KnownColor.MediumVioletRed, KnownColor.MidnightBlue, KnownColor.MintCream,
-                KnownColor.MistyRose, KnownColor.Moccasin, KnownColor.NavajoWhite, KnownColor.Navy, KnownColor.OldLace,
-                KnownColor.Olive, KnownColor.OliveDrab, KnownColor.Orange, KnownColor.OrangeRed, KnownColor.Orchid,
-                KnownColor.PaleGoldenrod, KnownColor.PaleGreen, KnownColor.PaleTurquoise, KnownColor.PaleVioletRed,
-                KnownColor.PapayaWhip, KnownColor.PeachPuff, KnownColor.Peru, KnownColor.Pink, KnownColor.Plum,
-                KnownColor.PowderBlue, KnownColor.Purple, KnownColor.Red, KnownColor.RosyBrown, KnownColor.RoyalBlue,
-                KnownColor.SaddleBrown, KnownColor.Salmon, KnownColor.SandyBrown, KnownColor.SeaGreen,
-                KnownColor.SeaShell, KnownColor.Sienna, KnownColor.Silver, KnownColor.SkyBlue, KnownColor.SlateBlue,
-                KnownColor.SlateGray, KnownColor.Snow, KnownColor.SpringGreen, KnownColor.SteelBlue, KnownColor.Tan,
-                KnownColor.Teal, KnownColor.Thistle, KnownColor.Tomato, KnownColor.Turquoise, KnownColor.Violet,
-                KnownColor.Wheat, KnownColor.White, KnownColor.WhiteSmoke, KnownColor.Yellow, KnownColor.YellowGreen
-            }.Select(kc => new object[] {kc}).ToArray();
-
         private Color? GetColorByProperty(string name)
         {
             return (Color?)typeof(Color).GetProperty(name)?.GetValue(null);
@@ -303,7 +247,6 @@ namespace System.Drawing.Primitives.Tests
 
         [Theory]
         [MemberData(nameof(ColorNamePairs))]
-        [InlineData("SomeUnknownColorName", "AnotherUnknownColorName")]
         public void GetHashCode(string name1, string name2)
         {
             Assert.NotEqual(name1, name2);
@@ -489,116 +432,6 @@ namespace System.Drawing.Primitives.Tests
         public void GetSaturation(int r, int g, int b, float expected)
         {
             Assert.Equal(expected, Color.FromArgb(r, g, b).GetSaturation());
-        }
-
-        [Theory, MemberData(nameof(NamedArgbValues))]
-        public void FromKnownColor(string name, int alpha, int red, int green, int blue)
-        {
-            Color color = Color.FromKnownColor(Enum.Parse<KnownColor>(name));
-            Assert.Equal(alpha, color.A);
-            Assert.Equal(red, color.R);
-            Assert.Equal(green, color.G);
-            Assert.Equal(blue, color.B);
-        }
-
-        [Fact]
-        public void FromOutOfRangeKnownColor()
-        {
-            Color color = Color.FromKnownColor((KnownColor)(-1));
-            Assert.Equal(0, color.A);
-            Assert.Equal(0, color.R);
-            Assert.Equal(0, color.G);
-            Assert.Equal(0, color.B);
-
-            color = Color.FromKnownColor(0);
-            Assert.Equal(0, color.A);
-            Assert.Equal(0, color.R);
-            Assert.Equal(0, color.G);
-            Assert.Equal(0, color.B);
-
-            color = Color.FromKnownColor(KnownColor.MenuHighlight + 1);
-            Assert.Equal(0, color.A);
-            Assert.Equal(0, color.R);
-            Assert.Equal(0, color.G);
-            Assert.Equal(0, color.B);
-        }
-
-        [Theory, MemberData(nameof(AllKnownColors))]
-        public void ToKnownColor(KnownColor known) => Assert.Equal(known, Color.FromKnownColor(known).ToKnownColor());
-
-        [Theory, MemberData(nameof(AllKnownColors))]
-        public void ToKnownColorMatchesButIsNotKnown(KnownColor known)
-        {
-            Color color = Color.FromKnownColor(known);
-            Color match = Color.FromArgb(color.A, color.R, color.G, color.B);
-            Assert.Equal((KnownColor)0, match.ToKnownColor());
-        }
-
-        [Fact]
-        public void FromOutOfRangeKnownColorToKnownColor()
-        {
-            Color color = Color.FromKnownColor((KnownColor)(-1));
-            Assert.Equal((KnownColor)0, color.ToKnownColor());
-
-            color = Color.FromKnownColor(0);
-            Assert.Equal((KnownColor)0, color.ToKnownColor());
-
-            color = Color.FromKnownColor(KnownColor.MenuHighlight + 1);
-            Assert.Equal((KnownColor)0, color.ToKnownColor());
-        }
-
-        [Theory, MemberData(nameof(SystemColors))]
-        public void IsSystemColorTrue(KnownColor known) => Assert.True(Color.FromKnownColor(known).IsSystemColor);
-
-        [Theory, MemberData(nameof(NonSystemColors))]
-        public void IsSystemColorFalse(KnownColor known) => Assert.False(Color.FromKnownColor(known).IsSystemColor);
-
-        [Theory, MemberData(nameof(SystemColors))]
-        public void IsSystemColorFalseOnMatching(KnownColor known)
-        {
-            Color color = Color.FromKnownColor(known);
-            Color match = Color.FromArgb(color.A, color.R, color.G, color.B);
-            Assert.False(match.IsSystemColor);
-        }
-
-        [Fact]
-        public void IsSystemColorOutOfRangeKnown()
-        {
-            Color color = Color.FromKnownColor((KnownColor)(-1));
-            Assert.False(color.IsSystemColor);
-
-            color = Color.FromKnownColor(0);
-            Assert.False(color.IsSystemColor);
-
-            color = Color.FromKnownColor(KnownColor.MenuHighlight + 1);
-            Assert.False(color.IsSystemColor);
-        }
-
-        [Theory, MemberData(nameof(AllKnownColors))]
-        public void IsKnownColorTrue(KnownColor known)
-        {
-            Assert.True(Color.FromKnownColor(known).IsKnownColor);
-        }
-
-        [Theory, MemberData(nameof(AllKnownColors))]
-        public void IsKnownColorMatchFalse(KnownColor known)
-        {
-            Color color = Color.FromKnownColor(known);
-            Color match = Color.FromArgb(color.A, color.R, color.G, color.B);
-            Assert.False(match.IsKnownColor);
-        }
-
-        [Fact]
-        public void IsKnownColorOutOfRangeKnown()
-        {
-            Color color = Color.FromKnownColor((KnownColor)(-1));
-            Assert.False(color.IsKnownColor);
-
-            color = Color.FromKnownColor(0);
-            Assert.False(color.IsKnownColor);
-
-            color = Color.FromKnownColor(KnownColor.MenuHighlight + 1);
-            Assert.False(color.IsKnownColor);
         }
 
         public static IEnumerable<object[]> Equality_MemberData()
